requirements.txt
openai-whisper
numpy==1.23.5
gensim==3.8.3
nltk
webrtcvad==2.0.10
hnswlib==0.7.0
sentence_transformers==2.2.2
protobuf==4.21.6
google-cloud-texttospeech
python-docx==0.8.11
beautifulsoup4==4.12.2